LLM-Voice-Assistant/server/requirements.txt
requests
playsound
faster_whisper
flask
langchain
langchain-community
langchain-core
piper-tts
